+++ src/main/java/com/example/quaterback/api/feature/monitoring/dto/response/DailyUsageDto.java
@@ -12,9 +12,9 @@ import lombok.NoArgsConstructor;
 @Builder
 public class DailyUsageDto {
     //일일정보
-    Long totalChargedEnergy;
+    Double totalChargedEnergy;
     Long totalVehicleCount;
-    Long totalRevenue;
+    Double totalRevenue;
 
     //퍼센트계산
     int chargedEnergyDiffPercent;
@@ -22,12 +22,12 @@ public class DailyUsageDto {
     int revenueDiffPercent;
 
     public static DailyUsageDto from(DailyUsageQuery today, DailyUsageQuery yesterday) {
-        long todayEnergy = today != null && today.getTotalChargedEnergy() != null ? today.getTotalChargedEnergy() : 0L;
-        long todayRevenue = today != null && today.getTotalRevenue() != null ? today.getTotalRevenue() : 0L;
+        double todayEnergy = today != null && today.getTotalChargedEnergy() != null ? today.getTotalChargedEnergy() : 0.0;
+        double todayRevenue = today != null && today.getTotalRevenue() != null ? today.getTotalRevenue() : 0.0;
         long todayCount = today != null && today.getTotalVehicleCount() != null ? today.getTotalVehicleCount() : 0;
 
-        long yesterdayEnergy = yesterday != null && yesterday.getTotalChargedEnergy() != null ? yesterday.getTotalChargedEnergy() : 0L;
-        long yesterdayRevenue = yesterday != null && yesterday.getTotalRevenue() != null ? yesterday.getTotalRevenue() : 0L;
+        double yesterdayEnergy = yesterday != null && yesterday.getTotalChargedEnergy() != null ? yesterday.getTotalChargedEnergy() : 0L;
+        double yesterdayRevenue = yesterday != null && yesterday.getTotalRevenue() != null ? yesterday.getTotalRevenue() : 0L;
         long yesterdayCount = yesterday != null && yesterday.getTotalVehicleCount() != null ? yesterday.getTotalVehicleCount() : 0;
 
         return DailyUsageDto.builder()
@@ -40,8 +40,8 @@ public class DailyUsageDto {
                 .build();
     }
 
-    private static int calculateDiffPercent(long today, long yesterday) {
+    private static int calculateDiffPercent(double today, double yesterday) {
         if (yesterday == 0) return today == 0 ? 0 : 100;
-        return (int) (((double) (today - yesterday) / yesterday) * 100);
+        return (int) (((today - yesterday) / yesterday) * 100);
     }
 }
